requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0.post1
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
beautifulsoup4==4.12.2
aiohttp==3.9.3
trafilatura==1.6.3
loguru==0.7.2
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.1.1
isort==5.13.2